First PlanetMapper changes
Updated sortIntoExtendedObservations to only have the 11 necessary processing files and have them be a single array with an obskey. Added process_L1X.py, which uses getL1AProcessingFiles to get a file list for initial processing by PlanetMapper software. The processing converts the PNG files to FITS files, both as images and projected maps and adds multiple 'backplanes' of metadata as a function of pixel location or latitude and longitude